ebook-generator/requirements.txt
# AST parsing
marko>=2.0.0

# Async HTTP
aiohttp>=3.9.0

# Image processing
pillow>=10.0.0

# Translation
google-generativeai>=0.3.0

# Environment
python-dotenv>=1.0.0
